blackout_lab/requirements.txt
Flask==3.1.3
PyJWT==2.12.1
lxml==6.0.2
PyYAML==6.0.3
gunicorn==21.2.0
PyMySQL==1.1.1
Flask-Dance==7.0.0
requests==2.31.0
